update api used in tutorials
