+++ doc/source/serve/tutorials/batch.md
@@ -18,9 +18,11 @@ This tutorial should help the following use cases:
   inference with batching can increase the *throughput* of the model as well as
   *utilization* of the hardware.
 
-Let's import Ray Serve and some other helpers.
 
-```{literalinclude} ../../../../python/ray/serve/examples/doc/tutorial_batch.py
+## Define the Deployment
+Open a new Python file called `tutorial_batch.py`. First, let's import Ray Serve and some other helpers.
+
+```{literalinclude} ../doc_code/tutorial_batch.py
 :end-before: __doc_import_end__
 :start-after: __doc_import_begin__
 ```
@@ -65,41 +67,92 @@ timeout may improve throughput at the cost of latency under low load.
 Let's define a deployment that takes in a list of requests, extracts the input value,
 converts them into an array, and uses NumPy to add 1 to each element.
 
-```{literalinclude} ../../../../python/ray/serve/examples/doc/tutorial_batch.py
+```{literalinclude} ../doc_code/tutorial_batch.py
 :end-before: __doc_define_servable_end__
 :start-after: __doc_define_servable_begin__
 ```
 
-Let's deploy it. Note that in the `@serve.batch` decorator, we are specifying
-the maximum batch size via `max_batch_size=4`. This option limits
+Let's prepare to deploy the deployment. Note that in the `@serve.batch` decorator, we
+are specifying the maximum batch size via `max_batch_size=4`. This option limits
 the maximum possible batch size that will be executed at once.
 
-```{literalinclude} ../../../../python/ray/serve/examples/doc/tutorial_batch.py
+```{literalinclude} ../doc_code/tutorial_batch.py
 :end-before: __doc_deploy_end__
 :start-after: __doc_deploy_begin__
 ```
 
-Let's define a [Ray remote task](ray-remote-functions) to send queries in
-parallel. As you can see, the first batch has a batch size of 1, and the subsequent
-queries have a batch size of 4. Even though each query is issued independently,
-Ray Serve was able to evaluate them in batches.
-
-```{literalinclude} ../../../../python/ray/serve/examples/doc/tutorial_batch.py
-:end-before: __doc_query_end__
-:start-after: __doc_query_begin__
+## Deploy the Deployment
+Deploy the deployment by running the following through the terminal.
+```console
+$ serve run tutorial_batch:adder
 ```
 
+Let's define a [Ray remote task](ray-remote-functions) to send queries in
+parallel. While Serve is running, open a separate terminal window, and run the 
+following in an interactive Python shell or a separate Python script:
+
+```python
+import ray
+import requests
+
+@ray.remote
+def send_query(number):
+    resp = requests.get("http://localhost:8000/?number={}".format(number))
+    return int(resp.text)
+
+
+# Let's use Ray to send all queries in parallel
+results = ray.get([send_query.remote(i) for i in range(9)])
+print("Result returned:", results)
+```
+
+You should get an output like the following. As you can see, the first batch has a 
+batch size of 1, and the subsequent queries have a batch size of 4. Even though each 
+query is issued independently, Ray Serve was able to evaluate them in batches.
+```bash
+(pid=...) Our input array has shape: (1,)
+(pid=...) Our input array has shape: (4,)
+(pid=...) Our input array has shape: (4,)
+Result returned: [1, 2, 3, 4, 5, 6, 7, 8, 9]
+```
+
+## Deploy the Deployment using Python API
 What if you want to evaluate a whole batch in Python? Ray Serve allows you to send
 queries via the Python API. A batch of queries can either come from the web server
-or the Python API. Learn more [here](serve-handle-explainer).
+or the Python API.
 
-To query the deployment via the Python API, we can use `Deployment.get_handle` to receive
-a handle to the corresponding deployment. To enqueue a query, you can call
-`handle.method.remote(data)`. This call returns immediately
-with a [Ray ObjectRef](ray-object-refs). You can call `ray.get` to retrieve
-the result.
+To query the deployment via the Python API, we can use `serve.run()`, which is part
+of the Python API, instead of running `serve run` from the console. Add the following
+to the Python script `tutorial_batch.py`:
 
-```{literalinclude} ../../../../python/ray/serve/examples/doc/tutorial_batch.py
-:end-before: __doc_query_handle_end__
-:start-after: __doc_query_handle_begin__
+```python
+handle = serve.run(adder)
+```
+
+Generally, to enqueue a query, you can call `handle.method.remote(data)`. This call 
+returns immediatelywith a [Ray ObjectRef](ray-object-refs). You can call `ray.get` to 
+retrieve the result. Add the following to the same Python script.
+
+```python
+input_batch = list(range(9))
+print("Input batch is", input_batch)
+
+result_batch = ray.get([handle.handle_batch.remote(i) for i in input_batch])
+print("Result batch is", result_batch)
+```
+
+Finally, let's run the script.
+```console
+$ python tutorial_batch.py
+```
+
+You should get an output like the following.
+```bash
+(pid=...) Current context is python
+(pid=...) Our input array has shape: (1,)
+(pid=...) Current context is python
+(pid=...) Our input array has shape: (4,)
+(pid=...) Current context is python
+(pid=...) Our input array has shape: (4,)
+Result batch is [1, 2, 3, 4, 5, 6, 7, 8, 9]
 ```
